Rewrite to less imperative style
Rubocop doesn't like the arg to `scrape` simply being called `h`. Rather
than simply renaming that to `config`, this is a good time to also bring
the scraper in-line with the more up-to-date approach of being less
imperative, and of making sure we can succeffully gather up all the data
before cleaning out the database.